remove NewFeatureAssertions, implement it directly in AnyDomain
we will make it configurable again with 0.10.0. Better we do not
introduce more ServiceLoader based assertions which we want to deprecate
with 0.10.0
