--- domain/robstoll/atrium-domain-robstoll-common/src/main/kotlin/ch/tutteli/atrium/domain/robstoll/creating/NewFeatureAssertionsImpl.kt
@@ -1,14 +1,0 @@
-@file:Suppress("DEPRECATION" /* TODO remove with 0.10.0 */)
-
-package ch.tutteli.atrium.domain.robstoll.creating
-
-import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.domain.creating.MetaFeature
-import ch.tutteli.atrium.domain.creating.NewFeatureAssertions
-import ch.tutteli.atrium.domain.robstoll.lib.creating._genericFeature
-
-class NewFeatureAssertionsImpl : NewFeatureAssertions {
-
-    override fun <T, R> genericFeature(assertionContainer: Expect<T>, metaFeature: MetaFeature<R>) =
-        _genericFeature(assertionContainer, metaFeature)
-}
