--- domain/robstoll/atrium-domain-robstoll-jvm/src/main/resources/META-INF/services/ch.tutteli.atrium.domain.creating.NewFeatureAssertions
@@ -1,1 +1,0 @@
-ch.tutteli.atrium.domain.robstoll.creating.NewFeatureAssertionsImpl
